Foreign.C.Types

Representations of C types

These types are needed to accurately represent C function prototypes, in order to access C library interfaces in Haskell. The Haskell system is not required to represent those types exactly as C does, but the following guarantees are provided concerning a Haskell type CT representing a C type t:

If a C function prototype has t as an argument or result type, the use of CT in the corresponding position in a foreign declaration permits the Haskell program to access the full range of values encoded by the C type; and conversely, any Haskell value for CT has a valid representation in C.

sizeOf (undefined :: CT) will yield the same value as sizeof (t) in C.

alignment (undefined :: CT) matches the alignment constraint enforced by the C implementation for t.

The members peek and poke of the Storable class map all values of CT to the corresponding value of t and vice versa.

When an instance of Bounded is defined for CT, the values of minBound and maxBound coincide with t_MIN and t_MAX in C.

When an instance of Eq or Ord is defined for CT, the predicates defined by the type class implement the same relation as the corresponding predicate in C on t.

When an instance of Num, Read, Integral, Fractional, Floating, RealFrac, or RealFloat is defined for CT, the arithmetic operations defined by the type class implement the same function as the corresponding arithmetic operations (if available) in C on t.

When an instance of Bits is defined for CT, the bitwise operation defined by the type class implement the same function as the corresponding bitwise operation in C on t.

Integral types

These types are are represented as newtypes of types in Data.Int and Data.Word, and are instances of Eq, Ord, Num, Read, Show, Enum, Storable, Bounded, Real, Integral and Bits.

data CChar

Haskell type representing the C char type.

data CSChar

Haskell type representing the C signed char type.

data CUChar

Haskell type representing the C unsigned char type.

data CShort

Haskell type representing the C short type.

data CUShort

Haskell type representing the C unsigned short type.

data CInt

Haskell type representing the C int type.

data CUInt

Haskell type representing the C unsigned int type.

data CLong

Haskell type representing the C long type.

data CULong

Haskell type representing the C unsigned long type.

data CPtrdiff

Haskell type representing the C ptrdiff_t type.

data CSize

Haskell type representing the C size_t type.

data CWchar

Haskell type representing the C wchar_t type.

data CSigAtomic

Haskell type representing the C sig_atomic_t type.

data CLLong

Haskell type representing the C long long type.

data CULLong

Haskell type representing the C unsigned long long type.

data CIntPtr

data CUIntPtr

data CIntMax

data CUIntMax

Numeric types

These types are are represented as newtypes of basic foreign types, and are instances of Eq, Ord, Num, Read, Show, Enum and Storable.

data CClock

Haskell type representing the C clock_t type.

data CTime

Haskell type representing the C time_t type.

Floating types

These types are are represented as newtypes of Float and Double, and are instances of Eq, Ord, Num, Read, Show, Enum, Storable, Real, Fractional, Floating, RealFrac and RealFloat.

data CFloat

Haskell type representing the C float type.

data CDouble

Haskell type representing the C double type.

Other types

data CFile

Haskell type representing the C FILE type.

data CFpos

Haskell type representing the C fpos_t type.

data CJmpBuf

Haskell type representing the C jmp_buf type.